"Remove unused plugin samples and add new feature
Unused sample code files and test files were removed to clean up the project. The changes have been made to focus on the main functionality of the plugin which is to show 'Hello World'. A new feature to show 'Hello World' as a notification was added. Plugin’s name and description were also updated in line with these changes."